@@ -1,3 +1,3 @@
 package com.jpro.routing
 
-case class Redirect(to: String) extends Result
+case class Redirect(to: String) extends Response
