@@ -1,5 +1,5 @@
 package com.jpro.routing.sessionmanager
-import com.jpro.routing.{Result, WebApp}
+import com.jpro.routing.{Response, WebApp}
 
 class DummySessionManager extends SessionManager {
   override def webApp: WebApp = null
@@ -8,9 +8,9 @@ class DummySessionManager extends SessionManager {
 
   override def goForward(): Unit = ()
 
-  override def gotoURL(_url: String, x: Result, pushState: Boolean, track: Boolean): Unit = ()
+  override def gotoURL(_url: String, x: Response, pushState: Boolean, track: Boolean): Unit = ()
 
-  override def getView(url: String): _root_.simplefx.experimental.FXFuture[Result] = null
+  override def getView(url: String): _root_.simplefx.experimental.FXFuture[Response] = null
 
   override def start(): Unit = ()
 }
